refactor(example): update example of using flutter camera plugin
